improvements and fixes
