+++ make_deb
@@ -78,11 +78,11 @@ BUILD_ROOT="${GIT_ROOT}/packages/Debian/build"
 package_name="foglamp-south-sinusoid-${version}-${architecture}"
 
 # Print the summary of findings
-echo "The package root directory is : ${GIT_ROOT}"
-echo "The FogLAMP version is        : ${version}"
-echo "The Package will be built in  : ${BUILD_ROOT}"
-echo "The architecture is set as    : ${architecture}"
-echo "The package name is           : ${package_name}"
+echo "The package root directory is         : ${GIT_ROOT}"
+echo "The FogLAMP south sinusoid version is : ${version}"
+echo "The Package will be built in          : ${BUILD_ROOT}"
+echo "The architecture is set as            : ${architecture}"
+echo "The package name is                   : ${package_name}"
 echo
 
 # Create the package directory. If a directory with the same name exists,
