+++ packages/Debian/armhf/DEBIAN/control
@@ -1,5 +1,5 @@
 Package: foglamp-south-sinusoid
-Version: 0.0
+Version: 1.0.0
 Section: devel
 Priority: optional
 Architecture: armhf
